Partial bytes/strings handling update.
In addition to a number of relatively mechanical changes (e[0] -> e.errno
or e.code, various string.join() syntax updates, adding "b" to mode for
relevant open() calls, and possibly a few others), which should be
comprehensive, this begins the work of updating string vs. byte objects
handling to be Python-3-compatible.

Basically, most if not all calls to xmlrpc.client.Binary(x) will need to
pass x.encode() instead, and references to item.data (which is the
retrieved version of the object that got stored by the Binary() call)
will need to become item.data.decode(). It's more complex than that, but
that will cover most cases.

What's not clear is whether adding those encoding/decoding calls will
mess up the previous author's use of base64 encoding to pass around data
without having to risk its getting mangled by passing through the wrong
character encoding.

I'm guessing it won't, because as I understand matters the Python 2
syntax was actually performing much the same encode/decode steps
implicitly, but I'm not entirely sure.

In any case, we'll need to be careful to leave data in base64 wherever
possible and only decode() it for external use (whether output, or
passing as a filepath to be opened, et cetera).

Large chunks of the codebase haven't been adjusted for this yet, but all
of the following moosic commands are working:

- play
- stop
- list
- history
- prepend
- no-adv