Fix unpausing with ogg123.
After ogg123 is sent a STOP signal, it appears that the next CONT signal
won't actually get it to start playing again - but the one after that
will.

There's already logic in skip() to give ogg123 special handling for the
TERM signal, so we just duplicate that under unpause() and adjust
accordingly. Seems solid in my testing so far.