Set myself as maintainer.
More specifically, update the author, author_email, and url fields in setup.py,
and add my name to README.txt.

Dropping references to the original author this way doesn't sit well with me,
but I don't know of any way to retain them in setup.py while still indicating
that current development responsibility goes to me, so here we are. With any
luck, the historical references to the original author in the codebase
(including README.txt) and the commit history will be enough.