Drop examples/completion.
We don't have a separate license statement for it, and the original
2003 copyright from Etienne PIERRE still applies, so Daniel Pearson's
change of moosic as a whole to the Unlicense doesn't apply to this
unless Etienne PIERRE agreed - and there's no sign that that happened.
Until such time as we can get the licensing clear, better to leave this
out.